+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Fixed an indentention issue (see #204).
 
+### Improved
+
+- Changed the location of the SM install command (see #202).
+
 ## [3.0.14]
 
 ### Added